Allow reviving with remaining teams just != 4

+++ app/controllers/concerns/revival.rb
@@ -5,36 +5,49 @@ module Revival
 	def revive_user(user)
 		eliminated = user.team.eliminated?
 		out_of_town = user.team.out_of_town?
-		
+
 		# Destroy all confirmed kills on the user.
 		user.kills.where(game_id: user.team.participation.game_id, confirmed: true).destroy_all
 
 		# Then revive the team if necessary.
 		if eliminated
-			# Destroy the old contract to eliminate this team.
-			Contract.where(target_id: user.team.id, completed: true).find { |contract| contract.participation.game_id == user.team.participation.game_id }.destroy
+			# If there is a contract order, preserve it.
+			if user.team.participation.game.remaining_teams.count > 5
+				# Destroy the old contract to eliminate this team.
+				Contract.where(target_id: user.team.id, completed: true).find { |contract| contract.participation.game_id == user.team.participation.game_id }.destroy
 
-			teams = view_context.contract_order_teams(user.team.participation.game, user.team.contract.next_non_eliminated_target)
+				teams = view_context.contract_order_teams(user.team.participation.game, user.team.contract.next_non_eliminated_target)
 
-			# Insert current team into last slot of contract order teams.
-			# Adjust last team's contract to be targeted at current team.
-			killer_contract = teams.last.contract
+				# Insert current team into last slot of contract order teams.
+				# Adjust last team's contract to be targeted at current team.
+				killer_contract = teams.last.contract
 
-			killer_contract.target_id = user.team.id
+				killer_contract.target_id = user.team.id
 
-			killer_contract.save
+				killer_contract.save
 
-			# Create contract for current team.
-			# Delete current contract.
-			user.team.contract.destroy
+				# Create contract for current team.
+				# Delete current contract.
+				user.team.contract.destroy
 
-			# Create new contract.
-			contract = Contract.new
-			contract.participation_id = user.team.participation.id
-			contract.target_id = teams.first.id
-			contract.start = DateTime.now
+				# Create new contract.
+				contract = Contract.new
+				contract.participation_id = user.team.participation.id
+				contract.target_id = teams.first.id
+				contract.start = DateTime.now
 
-			contract.save
+				contract.save
+			elsif user.team.participation.game.remaining_teams.count != 5
+				contracts = Contract.where(target_id: user.team.id, completed: true)
+				contracts.each do |c|
+					unless c.participation.team.eliminated?
+						c.end = nil
+						c.completed = false
+
+						c.save
+					end
+				end
+			end
 
 			# Reset termination_at for team.
 			participation = user.team.participation
@@ -52,7 +65,7 @@ module Revival
 		elsif out_of_town && !user.team.out_of_town?
 			# Get old kills.
 			kills = Kill.out_of_town.where(target_id: user.team.members.map { |member| member.id }, game_id: user.team.participation.game_id).where.not(appear_at: nil)
-			
+
 			# Adjust out_of_town_hours.
 			participation = user.team.participation
